Fix retries

+++ src/test/java/io/confluent/connect/elasticsearch/integration/ElasticsearchConnectorBaseIT.java
@@ -87,14 +87,15 @@ public class ElasticsearchConnectorBaseIT extends BaseConnectorIT {
   public void cleanup() throws IOException {
     stopConnect();
 
-    if (helperClient != null) {
-      try {
-        helperClient.deleteIndex(TOPIC);
-        helperClient.close();
-      } catch (ConnectException e) {
-        // Server is already down. No need to close
+    if (container.isRunning()) {
+      if (helperClient != null) {
+        try {
+          helperClient.deleteIndex(TOPIC);
+          helperClient.close();
+        } catch (ConnectException e) {
+          // Server is already down. No need to close
+        }
       }
-
     }
   }
